rag-pdf-llama-index/requirements.txt
gradio
gradio[mcp]
huggingface_hub[cli]
llama-cpp-python
llama-index
llama-index-readers-file
llama-index-vector-stores-postgres
llama-index-embeddings-huggingface
llama-index-llms-llama-cpp
psycopg2-binary
pgvector
asyncpg
greenlet
